Add function annotation to size_converter

+++ size_converter.py
@@ -1,13 +1,13 @@
-def sizeof_fmt_base2(num: int, suffix: str='B') -> str:
-    for unit in ['', 'Ki', 'Mi', 'Gi', 'Ti', 'Pi', 'Ei', 'Zi']:
+def sizeof_fmt_base2(num: float, suffix: str = "B") -> str:
+    for unit in ["", "Ki", "Mi", "Gi", "Ti", "Pi", "Ei", "Zi"]:
         if abs(num) < 1024.0:
             return f"{num:3.1f} {unit}{suffix}"
         num /= 1024.0
     return f"{num:.1f} Yi{suffix}"
 
 
-def sizeof_fmt_base10(num: int, suffix: str='B') -> str:
-    for unit in ['', 'K', 'M', 'G', 'T', 'P', 'E', 'Z']:
+def sizeof_fmt_base10(num: float, suffix: str = "B") -> str:
+    for unit in ["", "K", "M", "G", "T", "P", "E", "Z"]:
         if abs(num) < 1000.0:
             return f"{num:3.1f} {unit}{suffix}"
         num /= 1000.0
